+++ src/vm-repair/HISTORY.rst
@@ -1,10 +1,17 @@
 
 Release History
 ===============
-
 2.0.0
 ++++++
-Changed the default image for Windows VMs when the source image cannot be found to a Windows 2022 sku as opposed to Windows 2016.
+Changed default VM image to 2019-datacenter-gs for better default security. 
+
+1.1.0
+++++++
+Added script for GT fixit button
+
+1.0.10
+++++++
+Added breaking change warning for the default image for Windows source VMs if the source VM image is not found in `az vm repair create`. It will change from a 2016 image to 2022 in November 2024.
 
 1.0.9
 ++++++